debug

+++ dsbowl/modules/dataset.py
@@ -316,8 +316,8 @@ class MultiMasksList(SegmentationLabelList):
                     mask.numpy().squeeze().astype(np.uint8),
                     np.ones((3, 3),
                             np.uint8),
-                    iterations=1)).float().unsqueeze(0)
-        return ImageSegment(mask)
+                    iterations=1)).unsqueeze(0)
+        return ImageSegment(mask).long()
 
     def analyze_pred(self, pred, thresh: float = 0.5):
         return (pred > thresh).float()
